Update god_rays.comp
Test of new godrays, dual phased and phased outscatter fog. Current control is in the shader but before we had it coded in the constant buffer so I just have to go find the old build and move the code back in I just whated you to test it since we cant capture it 100% in video.