Switch to full-time OpenAL Soft
OpenAL mode now requires OpenAL soft 64 bit .dll to be present.

We've already beusing the dll
Fix the bug of the cinematics from 1.3 not stopping sound when you interrupt the cinematic, it keep playing sound even if you demo funmap1 it still plays then when its done playing the audio it throw user back to menu a slight crash, this is understanadble becuase how old the repo we are pulling from.