Great, another detour... ClientEntity won't stand a container from STL. It has its reasons, but what?
For this commit however, a partial move of SVG code to SG code has taken place. TypeInfo will be employed for both Client Game entities as well as Server Game entities. In the ideal circumstances, weapon code moves over to the SG folder and can be used for both game modules. Even more ideal but perhaps expecting too much, is to move the whole physics simulation over to the shared game code as well.