It's far from perfect yet, needs a lot of good looking at before moving on with implementations.
Nonetheless, it now spawns the client game class entities based on a server entity state's hashed classname.

func_ item_ etc are all hashed using the TypeInfo system. When hashes match on the client game side of things, it allocates the appropriate class entity to match with. If the hash is unknown to the client, it resorts to allocating a CLGBaseEntity which' purpose is to behave as if we never had these kind of entities to begin with.