Settled out on an adventurous day to introduce client/server-model lightstyle support for light emitting models.
Some minor fixes in how animations were dealt with for the current blending test scenario of our MonsterTestDummy.

Removed RenderEffects::MinimalLight for the gunRenderEntity. Seemed redundant as it seems to be a material tweak thing. Artists might wanna correct me here though.

Minor introduction of SharedGame items, to be continued.