cl_modellist now works properly. Useful to inspect which models are loaded by the client.
sv_modellist should work as expected. Useful to inspect which models have been loaded by the server. (Note that this is not done on the fly, but only per request to avoid unnescessary ram usage. As for why it says SHOULD, simple, it has only been tested with 1 model as of yet hehe)