Removed weird file Shared.cmakexx in /src/Shared.
Fixed Debris and Gibs to actually move into random directions after spawning. (Needs more perfecting of the art, but for now this'll do just fine.)

Renamed the server model API functions alike to those of the Client in the name of consistency.

SV_Physics_Toss is a piece of WIP. To be continued, right now it at least fixed(and was the cause of) the seeming appearance of our client entities becoming static after touching ground.

Added fmt library to start work on redoing some base code such as the prints which are very inconsistent right now. Time to clean that up.