WIP Commit: Progression and regression.
Improved what seemed to be a missing LERP bug for player weapon model.

Added EntitySkeleton functionality to easily access specific data from the Skeleton and its model data.

FuncPlat had a change of how it operates. To be continued.

Player Prediction, local clients and frame time in general is a WIP thing.